ENKI:  add new previewRefDelete repos API call; optimize mass-deletion via "indiscriminate" deletion of one-to-many links;
[git-p4: depot-paths = "//open/enki/": change = 11555]

+++ src/org/eigenbase/enki/hibernate/storage/HibernateObject.java
@@ -110,6 +110,10 @@ public abstract class HibernateObject
         }
         
         repos.checkTransaction(true);
+
+        if (repos.inPreviewDelete()) {
+            return;
+        }
         
         Session session = repos.getCurrentSession();
         session.delete(this);
